fix: dividing dynamic tests into test_actions.py file
