+++ tests/simulators/carla/test_basic.py
@@ -1,0 +1,51 @@
+import pytest
+
+from tests.utils import compileScenic, pickle_test, sampleScene, tryPickling
+
+# Suppress potential warning about missing the carla package
+pytestmark = pytest.mark.filterwarnings(
+    "ignore::scenic.core.simulators.SimulatorInterfaceWarning"
+)
+
+
+def test_map_param_parse(getAssetPath):
+    mapPath = getAssetPath("maps/CARLA/Town01.xodr")
+    code = f"""
+        param map = r'{mapPath}'
+        model scenic.simulators.carla.model
+        ego = new Car
+    """
+    scenario = compileScenic(code, mode2D=True)
+    assert scenario.params["carla_map"] == "Town01"
+
+
+def test_basic(loadLocalScenario):
+    scenario = loadLocalScenario("basic.scenic", mode2D=True)
+    scenario.generate(maxIterations=1000)
+
+
+def test_simulator_import():
+    pytest.importorskip("carla")
+    from scenic.simulators.carla import CarlaSimulator
+
+
+def test_consistent_object_type(getAssetPath):
+    pytest.importorskip("carla")
+    mapPath = getAssetPath("maps/CARLA/Town01.xodr")
+    code = f"""
+        param map = r'{mapPath}'
+        param carla_map = 'Town01'
+        model scenic.simulators.carla.model
+        action = SetGearAction(0)
+        ego = new Car
+        assert action.canBeTakenBy(ego)
+    """
+    for _ in range(2):
+        compileScenic(code, mode2D=True)
+
+
+@pickle_test
+@pytest.mark.slow
+def test_pickle(loadLocalScenario):
+    scenario = tryPickling(loadLocalScenario("basic.scenic", mode2D=True))
+    tryPickling(sampleScene(scenario, maxIterations=1000))
